#445 Manager command ‘show @@processor’ has the wrong result

+++ src/main/java/com/actiontech/dble/buffer/ByteBufferPage.java
@@ -104,6 +104,6 @@ public class ByteBufferPage {
     }
 
     public long getUsage() {
-        return chunkAllocateTrack.cardinality() * chunkSize;
+        return chunkAllocateTrack.cardinality() * (long) chunkSize;
     }
 }
